Adds env var tests for enabled and disabled with attributes
